edit methods in CategoryController and edit in view files

+++ app/Http/Controllers/Dashboard/CategoryController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers\Dashboard;
 use App\Http\Controllers\Controller;
 use App\Http\Requests\CategoryRequest;
 use App\Models\Category;
+use App\Models\Setting;
 use App\Models\User;
 use Illuminate\Http\Request;
 use Illuminate\Support\Str;
@@ -13,6 +14,11 @@ use Yajra\DataTables\DataTables;
 
 class CategoryController extends Controller
 {
+    protected $setting;
+    public function __construct(Setting $setting)
+    {
+        $this->setting = $setting;
+    }
     /**
      * Display a listing of the resource.
      */
@@ -23,27 +29,26 @@ class CategoryController extends Controller
 
     public function getCategoriesDatatable()
     {
-
         $data = Category::select('*')->with('parents');
-        return DataTables::of($data)
+        return  Datatables::of($data)
             ->addIndexColumn()
-            ->addColumn('action',function ($row){
-                return $btn = '
-                <a href="'.Route('dashboard.category.edit',$row->id).'" class="edit btn btn-success btn-sm"><i class="fa fa-edit"></i><a/>
-                <a id="deleteBtn" data-id="'.$row->id.'" class="edit btn btn-danger btn-sm" data-toggle="modal"
-                data-target="#deletemodal"><i class="fa fa-trash"></i></a>';
+            ->addColumn('action', function ($row) {
+                if (auth()->user()->can('viewAny', $this->setting)) {
+                    return $btn = '
+                        <a href="' . Route('dashboard.category.edit', $row->id) . '"  class="edit btn btn-success btn-sm" ><i class="fa fa-edit"></i></a>
+                        <a id="deleteBtn" data-id="' . $row->id . '" class="edit btn btn-danger btn-sm"  data-toggle="modal" data-target="#deletemodal"><i class="fa fa-trash"></i></a>';
+                }
             })
-            ->addColumn('parent',function ($row) {
-                return ($row->parent == 0) ? trans('words.main category') : $row->parents->translate(app()->getLocale())->title;
+            ->addColumn('parent', function ($row) {
+                return ($row->parent ==  0) ? trans('words.main category') :   $row->parents->translate(app()->getLocale())->title;
             })
-            ->addColumn('title',function ($row){
+            ->addColumn('title', function ($row) {
                 return $row->translate(app()->getLocale())->title;
             })
             ->addColumn('status', function ($row) {
                 return $row->status == null ? __('words.not activated') : __('words.' . $row->status);
             })
-//
-            ->rawColumns(['action', 'status' , 'title'])
+            ->rawColumns(['action', 'status', 'title'])
             ->make(true);
     }
 
@@ -52,6 +57,7 @@ class CategoryController extends Controller
      */
     public function create()
     {
+        $this->authorize('viewAny', $this->setting);
         $categories = Category::whereNull('parent')->orwhere('parent',0)->get();
         return view('dashboard.categories.add' , compact('categories'));
     }
@@ -61,7 +67,7 @@ class CategoryController extends Controller
      */
     public function store(CategoryRequest $request)
     {
-
+        $this->authorize('viewAny', $this->setting);
         $category = category::create($request->except('image', '_token'));
         if ($request->file(''))
         {
@@ -87,6 +93,7 @@ class CategoryController extends Controller
      */
     public function edit(Category $category)
     {
+        $this->authorize('viewAny', $this->setting);
         $categories = Category::whereNull('parent')->orWhere('parent', 0)->get();
         return view('dashboard.categories.edit', compact('category', 'categories'));
     }
@@ -96,6 +103,7 @@ class CategoryController extends Controller
      */
     public function update(Request $request, Category $category)
     {
+        $this->authorize('viewAny', $this->setting);
         $category->update($request->except('image', '_token'));
         if ($request->file('image')) {
             $file = $request->file('image');
@@ -117,6 +125,7 @@ class CategoryController extends Controller
     }
     public function delete(Request $request)
     {
+        $this->authorize('viewAny', $this->setting);
         if (is_numeric($request->id))
         {
             Category::where('parent',$request->id)->delete();
